requirements.txt
python-telegram-bot[job-queue]==20.6
requests
python-dotenv
quart
hypercorn
supabase
